Add actor_prelude_sample_part_name.xml and expression_test.xml

+++ src/lib/dataTypes/nonNegativeDecimal.ts
@@ -4,4 +4,4 @@ import { t } from '../schema';
  *
  * {@link https://www.w3.org/2021/06/musicxml40/musicxml-reference/data-types/non-negative-decimal/}
  */
-export const nonNegativeDecimal = () => t.int({ min: 0 });
+export const nonNegativeDecimal = () => t.float({ min: 0 });
